requirements.txt
Django==3.1.4
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
whitenoise==5.2.0
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0